I have written one more ant script, which enables jars building and some simple execution.
However jars do not contain 'name' sections in their manifests which means they can not be executed themselfs but only called forom enother class.

I havent learn yet, how to pass environment to the spawned JVM. Bekause it executes like daemon. It would be grate if any of you find the way to pass the descriptor '1' to it.